requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
yt-dlp==2024.8.6
PyNaCl==1.5.0
spotipy==2.24.0